requirements.txt
allure-pytest==2.13.2
fastapi==0.89.1
httpx==0.24.1
loguru==0.7.0
pydantic==1.10.11
pytest==7.4.0
pytest-cov==4.1.0
pytest-xdist==3.3.1
requests==2.28.2
uvicorn==0.21.1